fmpq_poly scalar
